Vinay: Device Usage Work Flow change
Vinay: Device Usage Work Flow change

+++ src/main/java/com/gv/midway/pojo/usageView/DeviceUsageViewElement.java
@@ -9,7 +9,7 @@ public class DeviceUsageViewElement {
 
     private DeviceId deviceId;
     private Integer netSuiteId;
-    private float dataUsed;
+    private long dataUsed;
     private String jobId;
     private Date lastTimeStampUpdated;
     private Boolean isUpdatedElement;
@@ -27,10 +27,10 @@ public class DeviceUsageViewElement {
     public void setNetSuiteId(Integer netSuiteId) {
         this.netSuiteId = netSuiteId;
     }
-    public float getDataUsed() {
+    public long getDataUsed() {
         return dataUsed;
     }
-    public void setDataUsed(float dataUsed) {
+    public void setDataUsed(long dataUsed) {
         this.dataUsed = dataUsed;
     }
     public String getJobId() {
